Loosen meeting attendance requirement
The current requirement is to stringent for our board population and
meeting frequency.  Assuming that we have 24 meetings per year a board
member could attend one meeting per month and be ejected from the board
for one unexpected illness.

Automatic expulsion is a good tool and reminder of expectations, but it
should not be difficult to avoid.